requirements.txt
flask
python-telegram-bot[webhooks]==20.7
google-generativeai
python-dotenv
asgiref
uvicorn
aiosqlite
